# openapi.yaml
openapi: 3.1.0
info:
  title: Documan API
  version: 1.0.0
  description: |
    REST API for organising academic subjects and files, and for the surrounding community
    features (posts, comments, votes, favourites, users, roles).

    This document is maintained by hand rather than generated at runtime. When an endpoint,
    payload or status code changes, update it in the same commit.

    Notes on conventions:
      * Collection endpoints are paginated and return a `PageResponse` envelope.
      * Errors use RFC 9457 `application/problem+json`.
      * Identifiers are passed as query parameters.
      * There is currently no authentication; every operation is reachable anonymously.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /api/v1/user:
    get:
      tags: [User]
      summary: Fetch a user by id
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "200": { $ref: "#/components/responses/User" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [User]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
    put:
      tags: [User]
      summary: Update a user
      description: A null `password` leaves the stored credential unchanged.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [User]
      summary: Delete a user
      description: Fails with 409 if the user still owns posts or comments.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/user/all:
    get:
      tags: [User]
      summary: List users
      parameters: [{ $ref: "#/components/parameters/Pageable" }]
      responses:
        "200": { $ref: "#/components/responses/UserPage" }
...
  /api/v1/subject:
    get:
      tags: [Subject]
      summary: Fetch a subject by id
      parameters: [{ $ref: "#/components/parameters/SubjectId" }]
      responses:
        "200": { $ref: "#/components/responses/Subject" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Subject]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubjectBody" }
      responses:
        "201": { $ref: "#/components/responses/Subject" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Subject]
      summary: Update a subject
      parameters: [{ $ref: "#/components/parameters/SubjectId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubjectBody" }
      responses:
        "200": { $ref: "#/components/responses/Subject" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Subject]
      summary: Delete a subject
      parameters: [{ $ref: "#/components/parameters/SubjectId" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/subject/all:
    get:
      tags: [Subject]
      summary: List subjects
      parameters: [{ $ref: "#/components/parameters/Pageable" }]
      responses:
        "200": { $ref: "#/components/responses/SubjectPage" }
  /api/v1/subject/semester:
    get:
      tags: [Subject]
      summary: List subjects for a department, year and semester
      parameters:
        - { $ref: "#/components/parameters/DepartmentId" }
        - { $ref: "#/components/parameters/YearId" }
        - { $ref: "#/components/parameters/SemesterId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/SubjectPage" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/role:
    get:
      tags: [Role]
      summary: Fetch a role by id
      parameters:
        - name: roleId
          in: query
          required: true
          schema: { type: integer, format: int32 }
      responses:
        "200": { $ref: "#/components/responses/Role" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/role/user:
    get:
      tags: [Role]
      summary: Fetch the role assigned to a user
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "200": { $ref: "#/components/responses/Role" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/role/promote:
    put:
      tags: [Role]
      summary: Move a user to a higher-ranked role
      description: Privilege rank is the numeric role id. Fails if the target does not rank above the current role.
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/RoleId" }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/role/demote:
    put:
      tags: [Role]
      summary: Move a user to a lower-ranked role
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/RoleId" }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/semester:
    get:
      tags: [Reference]
      summary: Fetch a semester by id
      parameters: [{ $ref: "#/components/parameters/SemesterId" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SemesterResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/search/files:
    get:
      tags: [Search]
      summary: Full-text search over files
      description: |
        Matches on filename, extension and the denormalised subject name and code. Returns `503`
        with `Retry-After` when Meilisearch is unavailable; the endpoint is absent entirely when
        `documan.search.enabled` is false.
      parameters:
        - { $ref: "#/components/parameters/SearchQuery" }
        - { name: subjectId, in: query, required: false, schema: { type: integer } }
        - { name: departmentId, in: query, required: false, schema: { type: integer } }
        - { name: yearId, in: query, required: false, schema: { type: integer } }
        - { name: semesterId, in: query, required: false, schema: { type: integer } }
        - { name: extension, in: query, required: false, schema: { type: string }, description: "Lower-case, without the dot." }
        - { name: lab, in: query, required: false, schema: { type: boolean } }
        - { name: theory, in: query, required: false, schema: { type: boolean } }
        - { $ref: "#/components/parameters/SearchPage" }
        - { $ref: "#/components/parameters/SearchSize" }
        - { $ref: "#/components/parameters/SearchSort" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/SearchResponse" }
                  - type: object
                    properties:
                      hits: { type: array, items: { $ref: "#/components/schemas/FileDocument" } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "503": { $ref: "#/components/responses/SearchUnavailable" }
  /api/v1/search/subjects:
    get:
      tags: [Search]
      summary: Full-text search over subjects
      parameters:
        - { $ref: "#/components/parameters/SearchQuery" }
        - { name: departmentId, in: query, required: false, schema: { type: integer } }
        - { name: yearId, in: query, required: false, schema: { type: integer } }
        - { name: semesterId, in: query, required: false, schema: { type: integer } }
        - { name: lab, in: query, required: false, schema: { type: boolean } }
        - { name: theory, in: query, required: false, schema: { type: boolean } }
        - { $ref: "#/components/parameters/SearchPage" }
        - { $ref: "#/components/parameters/SearchSize" }
        - { $ref: "#/components/parameters/SearchSort" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/SearchResponse" }
                  - type: object
                    properties:
                      hits: { type: array, items: { $ref: "#/components/schemas/SubjectDocument" } }
        "503": { $ref: "#/components/responses/SearchUnavailable" }
  /api/v1/search/posts:
    get:
      tags: [Search]
      summary: Full-text search over posts
      description: Matches on title, description, body and author name. Bodies are truncated in the index.
      parameters:
        - { $ref: "#/components/parameters/SearchQuery" }
        - { name: authorId, in: query, required: false, schema: { type: integer } }
        - { $ref: "#/components/parameters/SearchPage" }
        - { $ref: "#/components/parameters/SearchSize" }
        - { $ref: "#/components/parameters/SearchSort" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/SearchResponse" }
                  - type: object
                    properties:
                      hits: { type: array, items: { $ref: "#/components/schemas/PostDocument" } }
        "503": { $ref: "#/components/responses/SearchUnavailable" }
  /api/v1/search/comments:
    get:
      tags: [Search]
      summary: Full-text search over comments
      parameters:
        - { $ref: "#/components/parameters/SearchQuery" }
        - { name: postId, in: query, required: false, schema: { type: integer } }
        - { name: authorId, in: query, required: false, schema: { type: integer } }
        - { $ref: "#/components/parameters/SearchPage" }
        - { $ref: "#/components/parameters/SearchSize" }
        - { $ref: "#/components/parameters/SearchSort" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/SearchResponse" }
                  - type: object
                    properties:
                      hits: { type: array, items: { $ref: "#/components/schemas/CommentDocument" } }
        "503": { $ref: "#/components/responses/SearchUnavailable" }
components:
  parameters:
    SearchQuery:
      name: q
      in: query
      required: false
      description: Free-text query. Omit or leave blank to browse with filters only.
      schema: { type: string }
    SearchPage:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    SearchSize:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    SearchSort:
      name: sort
      in: query
      required: false
      description: "Meilisearch sort expression over a sortable attribute, e.g. `dateCreated:desc`."
      schema: { type: string }
    UserId:
      name: userId
      in: query
      required: true
      schema: { type: integer, format: int32 }
...
    SubjectId:
      name: subjectId
      in: query
      required: true
      schema: { type: integer, format: int32 }
...
    RoleId:
      name: roleId
      in: query
      required: true
      schema: { type: integer, format: int32 }
    DepartmentId:
      name: departmentId
      in: query
      required: true
      schema: { type: integer, format: int32 }
    YearId:
      name: yearId
      in: query
      required: true
      schema: { type: integer, format: int32 }
    SemesterId:
      name: semesterId
      in: query
      required: true
      schema: { type: integer, format: int32 }
...
    Pageable:
      name: pageable
      in: query
      required: false
      description: |
        Standard Spring Data pagination. Supply as separate query parameters:
        `page` (0-based), `size` (max 100), and `sort` as `property,asc|desc` (repeatable).
      style: form
      explode: true
      schema:
        type: object
        properties:
          page: { type: integer, minimum: 0, default: 0 }
          size: { type: integer, minimum: 1, maximum: 100, default: 20 }
          sort:
            type: array
            items: { type: string }
  responses:
    User:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UserResponse" }
    UserPage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/UserResponse" }
...
    Subject:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SubjectResponse" }
    SubjectPage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/SubjectResponse" }
...
    Role:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RoleResponse" }
    SearchUnavailable:
      description: Meilisearch could not be reached
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    NotFound:
      description: The addressed resource does not exist
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Conflict:
      description: Uniqueness, referential or optimistic-locking conflict
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    ValidationFailed:
      description: One or more fields failed validation
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ValidationProblemDetail" }
...
  schemas:
...
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last]
      properties:
        content: { type: array, items: {} }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
    CreateUserRequest:
      type: object
      required: [username, password, firstName, lastName, email, departmentId, yearId, semesterId]
      properties:
        username: { type: string, maxLength: 40 }
        password: { type: string, minLength: 8, maxLength: 200 }
        firstName: { type: string, maxLength: 60 }
        lastName: { type: string, maxLength: 60 }
        email: { type: string, format: email, maxLength: 180 }
        departmentId: { type: integer, format: int32 }
        yearId: { type: integer, format: int32 }
        semesterId: { type: integer, format: int32 }
        acceptedTermsOfService: { type: boolean, default: false }
    UpdateUserRequest:
      type: object
      required: [username, firstName, lastName, email, departmentId, yearId, semesterId]
      properties:
        username: { type: string, maxLength: 40 }
        password:
          type: [string, "null"]
          minLength: 8
          maxLength: 200
          description: Omit or send null to leave the stored password unchanged.
        firstName: { type: string, maxLength: 60 }
        lastName: { type: string, maxLength: 60 }
        email: { type: string, format: email, maxLength: 180 }
        departmentId: { type: integer, format: int32 }
        yearId: { type: integer, format: int32 }
        semesterId: { type: integer, format: int32 }
        acceptedTermsOfService: { type: [boolean, "null"] }
        canPost: { type: [boolean, "null"] }
        canComment: { type: [boolean, "null"] }
...
    SubjectBody:
      type: object
      required: [name, code, departmentId, yearId, semesterId]
      properties:
        name: { type: string, maxLength: 150 }
        code: { type: string, maxLength: 20 }
        lab: { type: boolean, default: false }
        theory: { type: boolean, default: false }
        departmentId: { type: integer, format: int32 }
        yearId: { type: integer, format: int32 }
        semesterId: { type: integer, format: int32 }
    DepartmentResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
    YearResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        value: { type: string }
    SemesterResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
    RoleResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
    UserResponse:
      type: object
      description: The password is never included in a response.
      properties:
        id: { type: integer, format: int32 }
        username: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        acceptedTermsOfService: { type: boolean }
        verified: { type: boolean }
        canPost: { type: boolean }
        canComment: { type: boolean }
        role: { $ref: "#/components/schemas/RoleResponse" }
        department: { $ref: "#/components/schemas/DepartmentResponse" }
        year: { $ref: "#/components/schemas/YearResponse" }
        semester: { $ref: "#/components/schemas/SemesterResponse" }
        dateCreated: { type: string, format: date-time }
        dateLastInteracted: { type: string, format: date-time }
...
    SubjectResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
        code: { type: string }
        lab: { type: boolean }
        theory: { type: boolean }
        department: { $ref: "#/components/schemas/DepartmentResponse" }
        year: { $ref: "#/components/schemas/YearResponse" }
        semester: { $ref: "#/components/schemas/SemesterResponse" }
...
    SearchResponse:
      type: object
      description: |
        `estimatedTotalHits` is Meilisearch's own wording: the engine trades an exact count for
        speed, so it must not be treated as authoritative for pagination.
      properties:
        hits: { type: array, items: {} }
        query: { type: string }
        page: { type: integer }
        size: { type: integer }
        estimatedTotalHits: { type: integer, format: int64 }
        processingTimeMs: { type: integer, format: int64 }
    FileDocument:
      type: object
      description: Timestamps are epoch seconds so ranges and sorting work numerically.
      properties:
        id: { type: integer }
        name: { type: string }
        extension: { type: string }
        objectName: { type: string }
        objectUrl: { type: string }
        size: { type: integer, format: int64 }
        favouriteCount: { type: integer, format: int64 }
        subjectId: { type: integer }
        subjectName: { type: string }
        subjectCode: { type: string }
        lab: { type: boolean }
        theory: { type: boolean }
        departmentId: { type: integer }
        departmentName: { type: string }
        yearId: { type: integer }
        yearValue: { type: string }
        semesterId: { type: integer }
        semesterName: { type: string }
        dateCreated: { type: integer, format: int64 }
    SubjectDocument:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        lab: { type: boolean }
        theory: { type: boolean }
        departmentId: { type: integer }
        departmentName: { type: string }
        yearId: { type: integer }
        yearValue: { type: string }
        semesterId: { type: integer }
        semesterName: { type: string }
    PostDocument:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        content: { type: string, description: Truncated; the full body comes from the post endpoint. }
        authorId: { type: integer }
        authorUsername: { type: string }
        upvoteCount: { type: integer, format: int64 }
        downvoteCount: { type: integer, format: int64 }
        favouriteCount: { type: integer, format: int64 }
        netScore: { type: integer, format: int64 }
        dateCreated: { type: integer, format: int64 }
        dateModified: { type: integer, format: int64 }
    CommentDocument:
      type: object
      properties:
        id: { type: integer }
        content: { type: string }
        postId: { type: integer }
        authorId: { type: integer }
        authorUsername: { type: string }
        upvoteCount: { type: integer, format: int64 }
        downvoteCount: { type: integer, format: int64 }
        dateCreated: { type: integer, format: int64 }
    ProblemDetail:
      type: object
      description: RFC 9457 problem detail.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        timestamp: { type: string, format: date-time }
    ValidationProblemDetail:
      allOf:
        - { $ref: "#/components/schemas/ProblemDetail" }
        - type: object
          properties:
            errors:
              type: object
              description: Field name to failure message.
              additionalProperties: { type: string }